pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scorpio-pro"
version = "1.0.0"
description = "State-of-the-art penetration testing and security auditing tool"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "jsdosanj" }]
requires-python = ">=3.11"
keywords = ["penetration-testing", "security", "auditing", "compliance", "vulnerability-scanning"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "click>=8.1.0",
    "psutil>=5.9.0",
    "python-nmap>=0.7.1",
    "paramiko>=3.0.0",
    "boto3>=1.26.0",
    "azure-identity>=1.12.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-network>=25.0.0",
    "azure-mgmt-storage>=21.0.0",
    "google-cloud-storage>=2.7.0",
    "google-cloud-resource-manager>=1.8.0",
    "google-cloud-iam>=2.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "cryptography>=39.0.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
]

[project.optional-dependencies]
scapy = ["scapy>=2.5.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
scorpio-pro = "scorpio_pro.cli:main"

[project.urls]
Homepage = "https://github.com/jsdosanj/scorpio-pro"
Repository = "https://github.com/jsdosanj/scorpio-pro"
Issues = "https://github.com/jsdosanj/scorpio-pro/issues"

[tool.setuptools.packages.find]
include = ["scorpio_pro*"]

[tool.setuptools.package-data]
"scorpio_pro.reporting.templates" = ["*.j2", "*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "S", "BLE", "A"]
ignore = ["S101", "S603", "S607"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
